+++ resources/views/tailwind/includes/table.blade.php
@@ -30,62 +30,7 @@
     </x-slot>
 
     <x-slot name="body">
-        @if (count($bulkActions) && (($selectPage && $rows->total() > $rows->count()) || count($selected)))
-            <x-livewire-tables::table.row wire:key="row-message" class="bg-indigo-50">
-                <x-livewire-tables::table.cell :colspan="count($bulkActions) ? count($columns) + 1 : count($columns)">
-                    @if (count($selected) && !$selectAll && !$selectPage)
-                        <div>
-                            <span>
-                                @lang('You have selected')
-                                <strong>{{ count($selected) }}</strong>
-                                @lang(':rows', ['rows' => count($selected) === 1 ? 'row' : 'rows']).
-                            </span>
-
-                            <button
-                                wire:click="resetBulk"
-                                type="button"
-                                class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
-                            >
-                                @lang('Unselect All')
-                            </button>
-                        </div>
-                    @elseif ($selectAll)
-                        <div>
-                            <span>
-                                @lang('You are currently selecting all')
-                                <strong>{{ number_format($rows->total()) }}</strong>
-                                @lang('rows').
-                            </span>
-
-                            <button
-                                wire:click="resetBulk"
-                                type="button"
-                                class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
-                            >
-                                @lang('Unselect All')
-                            </button>
-                        </div>
-                    @else
-                        <div>
-                            <span>
-                                @lang('You have selected')
-                                <strong>{{ $rows->count() }}</strong>
-                                @lang('rows, do you want to select all')
-                                <strong>{{ number_format($rows->total()) }}</strong>?
-                            </span>
-
-                            <button
-                                wire:click="selectAll"
-                                type="button"
-                                class="ml-1 text-blue-600 underline text-gray-700 text-sm leading-5 font-medium focus:outline-none focus:text-gray-800 focus:underline transition duration-150 ease-in-out"
-                            >
-                                @lang('Select All')
-                            </button>
-                        </div>
-                    @endif
-                </x-livewire-tables::table.cell>
-            </x-livewire-tables::table.row>
-        @endif
+        @include('livewire-tables::tailwind.includes.bulk-select-row')
 
         @forelse ($rows as $index => $row)
             <x-livewire-tables::table.row